// typescript/interfaces.ts
interface VendorData {
  firstName: string;
  lastName: string;
  profilePicture: string;
  type: "individual" | "company" | string;
  uid: string;
  _id: string;
}
// ...
interface CancellationPolicy {
  type: string;
  description: string;
}
// ...
export interface ServiceDataTwo {
  _id: string;
  user_id: string;
  vendor_data: VendorData;
  category_title: string;
  packages_data: PackageData[];
  title: string;
  packages: string[];
  description: string;
  category: string;
  sub_category: string;
  images: string[];
  address: string;
  country: string;
  city: string;
  state: string;
  zip_code: string;
  excluded_areas: string[];
  operating_area: string;
  booking_type: string;
  rules: any[];
  cancellation_policy: CancellationPolicy;
  currency: string;
  createdAt: string;
  updatedAt: string;
  status: string;
  isDeleted: boolean;
}

interface VendorData {
  _id: string;
  uid: string;
  firstName: string;
  lastName: string;
  profilePicture: string;
  type: string;
}

interface PackageData {
  _id: string;
  name: string;
  description: string;
  duration: string;
  features: string[];
  rate: number;
}